+++ src/test/php/net/xp_framework/unittest/reflection/MethodInvocationTest.class.php
@@ -73,5 +73,4 @@ class MethodInvocationTest extends MethodsTest {
     $fixture= $this->type($declaration);
     $this->assertEquals('Test', $fixture->getMethod('fixture')->setAccessible(true)->invoke($fixture->newInstance(), []));
   }
-
 }